migrated crew document categories

+++ lib/pages/admin/staff/actions.dart
@@ -1,4 +1,5 @@
 import 'package:adsats_amplify_gen_2/API/mutations.dart';
+import 'package:adsats_amplify_gen_2/helper/confirm_dialog.dart';
 import 'package:adsats_amplify_gen_2/models/ModelProvider.dart';
 import 'package:adsats_amplify_gen_2/pages/admin/staff/api.dart';
 import 'package:adsats_amplify_gen_2/pages/admin/staff/repo.dart';
@@ -32,23 +33,43 @@ class StaffActions extends ConsumerWidget {
         ),
         IconButton(
           onPressed: () async {
-            await Future.wait([
-              update(staff.copyWith(archived: !staff.archived)),
-              staff.archived ? enableUser(staff.id) : disableUser(staff.id),
-            ]);
-            ref.invalidate(staffRepoProvider);
-            controller.close();
+            final result = await showConfirmDialog(
+              context,
+              Text("Are you sure?"),
+              Text(
+                "Do you want to ${staff.archived ? "unarchive" : "archive"} this staff?",
+              ),
+            );
+            if (result) {
+              await Future.wait([
+                update(staff.copyWith(archived: !staff.archived)),
+                staff.archived ? enableUser(staff.id) : disableUser(staff.id),
+              ]);
+              ref.invalidate(staffRepoProvider);
+              controller.close();
+            }
           },
-          icon: const Icon(Icons.archive_outlined),
+          icon: Icon(
+            staff.archived ? Icons.unarchive_outlined : Icons.archive_outlined,
+          ),
+          tooltip:
+              staff.archived ? "Unarchive this staff" : "Archive this staff",
         ),
         IconButton(
           onPressed: () async {
-            await Future.wait([
-              deleteStaff(staff),
-              deleteUser(staff.id),
-            ]);
-            ref.invalidate(staffRepoProvider);
-            controller.close();
+            final result = await showConfirmDialog(
+              context,
+              Text("Are you sure?"),
+              Text("Do you want to delete this staff?"),
+            );
+            if (result) {
+              await Future.wait([
+                deleteStaff(staff),
+                deleteUser(staff.id),
+              ]);
+              ref.invalidate(staffRepoProvider);
+              controller.close();
+            }
           },
           icon: const Icon(Icons.delete_outline),
         ),
